Made some adjustments to Tuples.

+++ src/main/java/com/jeffreymanzione/jef/resurrection/JEFEntityTuple.java
@@ -103,7 +103,8 @@ public abstract class JEFEntityTuple extends JEFEntity<Integer> {
     String result;
     result = "(";
     try {
-      result += JEFEntity.getValueFromObject(this.get(0), indents, useSpaces, spacesPerTab);
+      result += JEFEntity.getValueFromObject(this.get((Integer) 0), indents, useSpaces,
+          spacesPerTab);
     } catch (CouldNotUpdateEntityException e) {
       result += "!!!";
       e.printStackTrace();
@@ -111,7 +112,8 @@ public abstract class JEFEntityTuple extends JEFEntity<Integer> {
     for (int i = 1; i < size(); i++) {
       result += ", ";
       try {
-        result += JEFEntity.getValueFromObject(this.get(i), indents, useSpaces, spacesPerTab);
+        result += JEFEntity.getValueFromObject(this.get((Integer) i), indents, useSpaces,
+            spacesPerTab);
       } catch (CouldNotUpdateEntityException e) {
         result += "!!!";
         e.printStackTrace();
@@ -122,5 +124,4 @@ public abstract class JEFEntityTuple extends JEFEntity<Integer> {
     return result;
 
   }
-
 }
